requirements.txt
tensorboardX==1.9
pytorch-lightning==1.5.10
pytorch-ignite>=0.4
scipy
test-tube>=0.7.1
tqdm>=4.41
fire>=0.2.1
opencv-python>=4.5
scikit-learn>=0.24.1
wandb>=0.10.31
segmentation-models-pytorch==0.2.0
pretrainedmodels==0.7.4
torchvision
torch
torchaudio
jupyterlab
notebook
matplotlib
portalocker